refactor: sidebar
