+++ app/src/features/apiClient/screens/apiClient/components/sidebar/APIClientSidebar.tsx
@@ -1,6 +1,13 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { RQAPI } from "../../../../types";
-import { ApiClientSidebarTabs } from "./apiClientSidebarTabs/ApiClientSidebarTabs";
+import { NavLink, useLocation } from "react-router-dom";
+import PATHS from "config/constants/sub/paths";
+import { Tabs, TabsProps, Tooltip } from "antd";
+import { BsCollection } from "@react-icons/all-files/bs/BsCollection";
+import { MdOutlineHistory } from "@react-icons/all-files/md/MdOutlineHistory";
+import { CollectionsList } from "./collectionsList/CollectionsList";
+import { HistoryList } from "./historyList/HistoryList";
+import { ApiClientSidebarHeader } from "./apiClientSidebarHeader/ApiClientSidebarHeader";
 import "./apiClientSidebar.scss";
 
 interface Props {
@@ -11,6 +18,11 @@ interface Props {
   onImportClick: () => void;
 }
 
+export enum ApiClientSidebarTabKey {
+  HISTORY = "history",
+  COLLECTIONS = "collections",
+}
+
 const APIClientSidebar: React.FC<Props> = ({
   history,
   onSelectionFromHistory,
@@ -18,14 +30,78 @@ const APIClientSidebar: React.FC<Props> = ({
   onNewClick,
   onImportClick,
 }) => {
+  const location = useLocation();
+  const [activeKey, setActiveKey] = useState<ApiClientSidebarTabKey>(ApiClientSidebarTabKey.COLLECTIONS);
+
+  useEffect(() => {
+    switch (location.pathname) {
+      case PATHS.API_CLIENT.ABSOLUTE: {
+        setActiveKey(ApiClientSidebarTabKey.COLLECTIONS);
+        return;
+      }
+      case PATHS.API_CLIENT.HISTORY.ABSOLUTE: {
+        setActiveKey(ApiClientSidebarTabKey.HISTORY);
+        return;
+      }
+      default: {
+        setActiveKey(ApiClientSidebarTabKey.COLLECTIONS);
+        return;
+      }
+    }
+  }, [location.pathname]);
+
+  const items: TabsProps["items"] = [
+    {
+      key: ApiClientSidebarTabKey.COLLECTIONS,
+      label: (
+        <Tooltip title="Collections" placement="right">
+          <NavLink
+            to={PATHS.API_CLIENT.ABSOLUTE}
+            className={({ isActive }) => `${isActive ? "active" : ""} api-client-tab-link`}
+          >
+            <BsCollection />
+          </NavLink>
+        </Tooltip>
+      ),
+      children: <CollectionsList onNewClick={onNewClick} onImportClick={onImportClick} />,
+    },
+    {
+      key: ApiClientSidebarTabKey.HISTORY,
+      label: (
+        <Tooltip title="History" placement="right">
+          <NavLink
+            to={PATHS.API_CLIENT.HISTORY.ABSOLUTE}
+            className={({ isActive }) => `${isActive ? "active" : ""} api-client-tab-link`}
+          >
+            <MdOutlineHistory />
+          </NavLink>
+        </Tooltip>
+      ),
+      children: <HistoryList history={history} onSelectionFromHistory={onSelectionFromHistory} />,
+    },
+  ];
+
+  const handleActiveTabChange = (activeKey: ApiClientSidebarTabKey) => {
+    setActiveKey(activeKey);
+  };
+
   return (
     <div className="api-client-sidebar">
-      <ApiClientSidebarTabs
+      <ApiClientSidebarHeader
+        activeTab={activeKey}
         history={history}
-        onSelectionFromHistory={onSelectionFromHistory}
-        clearHistory={clearHistory}
+        onClearHistory={clearHistory}
         onNewClick={onNewClick}
         onImportClick={onImportClick}
+      />
+
+      <Tabs
+        items={items}
+        size="small"
+        tabPosition="left"
+        className="api-client-sidebar-tabs"
+        activeKey={activeKey}
+        onChange={handleActiveTabChange}
       />
     </div>
   );
