+++ app/src/features/apiClient/screens/apiClient/components/sidebar/apiClientSidebar.scss
@@ -5,21 +5,6 @@
   flex-direction: column;
   height: 100%;
 
-  .api-client-sidebar-placeholder {
-    height: 100%;
-    width: 100%;
-    display: flex;
-    gap: 10px;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    padding: 8px;
-
-    .ant-typography {
-      text-align: center;
-    }
-  }
-
   .api-client-sidebar-header {
     flex: 0 0 auto;
     border-bottom: 1px solid var(--border-dark);
@@ -30,70 +15,156 @@
     margin-bottom: 8px;
   }
 
-  .api-history-list {
-    width: 100%;
+  .api-client-sidebar-tabs {
+    flex-shrink: 0;
     height: 100%;
-    flex: auto;
-    overflow: auto;
-    padding-left: 16px;
-    padding-top: 16px;
+    overflow: hidden;
+    background-color: var(--body-background);
 
-    .ant-timeline-item {
-      padding-bottom: 18px;
+    .api-client-sidebar-placeholder {
+      height: 100%;
+      width: 100%;
+      display: flex;
+      gap: 10px;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+      padding: 8px;
 
-      .ant-timeline-item-head {
-        margin-top: 3px;
-      }
-
-      .ant-timeline-item-tail {
-        margin-top: 12px;
-      }
-
-      .ant-timeline-item-content {
-        margin-left: 10px;
+      .ant-typography {
+        text-align: center;
       }
     }
 
-    .api-history-row {
-      display: flex;
-      align-items: baseline;
-      gap: 10px;
-      line-height: 23px;
-      padding: 4px;
-      margin: -8px 10px;
-      border-radius: 4px;
+    .api-history-list {
+      width: 100%;
+      height: 100%;
+      flex: auto;
+      overflow: auto;
+      padding-left: 16px;
+      padding-top: 16px;
 
-      &.clickable:hover {
-        cursor: pointer;
-        background-color: var(--gray-dark);
+      .ant-timeline-item {
+        padding-bottom: 18px;
 
-        .api-history-url {
-          color: var(--requestly-color-text-default);
+        .ant-timeline-item-head {
+          margin-top: 3px;
+        }
+
+        .ant-timeline-item-tail {
+          margin-top: 12px;
+        }
+
+        .ant-timeline-item-content {
+          margin-left: 10px;
         }
       }
 
-      .api-method {
-        flex-shrink: 0;
-        vertical-align: middle;
-        font-size: var(--requestly-font-size-2xs, 11px);
+      .api-history-row {
+        display: flex;
+        align-items: baseline;
+        gap: 10px;
+        line-height: 23px;
+        padding: 4px;
+        margin: -8px 10px;
+        border-radius: 4px;
+
+        &.clickable:hover {
+          cursor: pointer;
+          background-color: var(--gray-dark);
+
+          .api-history-url {
+            color: var(--requestly-color-text-default);
+          }
+        }
+
+        .api-method {
+          flex-shrink: 0;
+          vertical-align: middle;
+          font-size: var(--requestly-font-size-2xs, 11px);
+        }
+
+        .api-history-url {
+          font-size: var(--requestly-font-size-sm, 13px);
+          width: 100%;
+          overflow: hidden;
+          text-overflow: ellipsis;
+          word-break: break-all;
+          line-height: 18px;
+          display: -webkit-box;
+          line-clamp: 2;
+          -webkit-box-orient: vertical;
+          -webkit-line-clamp: 2;
+        }
+
+        .api-history-start-marker {
+          font-size: var(--requestly-font-size-sm, 13px);
+        }
+      }
+    }
+
+    .ant-tabs-nav {
+      margin-bottom: 0;
+      padding: 0;
+      min-width: 40px;
+
+      .ant-tabs-nav-list {
+        padding: 4px;
+        display: flex;
+        align-items: center;
+        gap: 8px;
       }
 
-      .api-history-url {
-        font-size: var(--requestly-font-size-sm, 13px);
-        width: 100%;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        word-break: break-all;
-        line-height: 18px;
-        display: -webkit-box;
-        line-clamp: 2;
-        -webkit-box-orient: vertical;
-        -webkit-line-clamp: 2;
-      }
+      .ant-tabs-tab {
+        margin: 0;
+        padding: 0;
+        border-radius: 4px;
 
-      .api-history-start-marker {
-        font-size: var(--requestly-font-size-sm, 13px);
+        &:hover,
+        &.ant-tabs-tab-active {
+          background-color: var(--requestly-color-surface-1);
+        }
+
+        .ant-tabs-tab-btn,
+        .ant-tabs-tab-btn:active {
+          color: var(--requestly-color-text-subtle);
+        }
+
+        .api-client-tab-link {
+          width: 32px;
+          height: 32px;
+          padding: 0 8px;
+          border-radius: 4px;
+
+          display: flex;
+          align-items: center;
+          gap: 4px;
+          font-size: var(--requestly-font-size-sm);
+          color: inherit;
+
+          svg {
+            width: 16px;
+            height: 16px;
+          }
+
+          &:hover {
+            background-color: var(--requestly-color-surface-1);
+          }
+        }
       }
+    }
+
+    .ant-tabs-content {
+      height: 100%;
+
+      .ant-tabs-tabpane {
+        padding: 0;
+        height: 100%;
+      }
+    }
+
+    .ant-tabs-ink-bar {
+      display: none;
     }
   }
 }
