--- app/src/features/apiClient/screens/apiClient/components/sidebar/apiClientSidebarTabs/apiClientSidebarTabs.scss
@@ -1,72 +1,0 @@
-.api-client-sidebar {
-  .api-client-sidebar-tabs {
-    flex-shrink: 0;
-    height: 100%;
-    overflow: hidden;
-    background-color: var(--body-background);
-
-    .ant-tabs-nav {
-      margin-bottom: 0;
-      padding: 0;
-      min-width: 40px;
-
-      .ant-tabs-nav-list {
-        padding: 4px;
-        display: flex;
-        align-items: center;
-        gap: 8px;
-      }
-
-      .ant-tabs-tab {
-        margin: 0;
-        padding: 0;
-        border-radius: 4px;
-
-        &:hover,
-        &.ant-tabs-tab-active {
-          background-color: var(--requestly-color-surface-1);
-        }
-
-        .ant-tabs-tab-btn,
-        .ant-tabs-tab-btn:active {
-          color: var(--requestly-color-text-subtle);
-        }
-
-        .api-client-tab-link {
-          width: 32px;
-          height: 32px;
-          padding: 0 8px;
-          border-radius: 4px;
-
-          display: flex;
-          align-items: center;
-          gap: 4px;
-          font-size: var(--requestly-font-size-sm);
-          color: inherit;
-
-          svg {
-            width: 16px;
-            height: 16px;
-          }
-
-          &:hover {
-            background-color: var(--requestly-color-surface-1);
-          }
-        }
-      }
-    }
-
-    .ant-tabs-content {
-      height: 100%;
-
-      .ant-tabs-tabpane {
-        padding: 0;
-        height: 100%;
-      }
-    }
-
-    .ant-tabs-ink-bar {
-      display: none;
-    }
-  }
-}
