+++ app/src/features/apiClient/screens/apiClient/components/sidebar/collectionsList/collectionsList.scss
