+++ app/src/features/apiClient/screens/apiClient/components/sidebar/historyList/HistoryList.tsx
@@ -1,10 +1,8 @@
 import React, { useCallback } from "react";
-import placeholderImage from "../../../../../../../../assets/images/illustrations/empty-sheets-dark.svg";
+import placeholderImage from "../../../../../../../assets/images/illustrations/empty-sheets-dark.svg";
 import { RQAPI } from "features/apiClient/types";
-import { ClearOutlined } from "@ant-design/icons";
-import { RQButton } from "lib/design-system-v2/components";
 import { Timeline, Typography } from "antd";
-import { REQUEST_METHOD_COLORS } from "../../../../../../../../constants";
+import { REQUEST_METHOD_COLORS } from "../../../../../../../constants";
 import { trackRequestSelectedFromHistory } from "modules/analytics/events/features/apiClient";
 import { trackRQDesktopLastActivity, trackRQLastActivity } from "utils/AnalyticsUtils";
 import { API_CLIENT } from "modules/analytics/events/features/constants";
@@ -12,10 +10,9 @@ import { API_CLIENT } from "modules/analytics/events/features/constants";
 interface Props {
   history: RQAPI.Entry[];
   onSelectionFromHistory: (index: number) => void;
-  onClearHistory: () => void;
 }
 
-export const HistoryList: React.FC<Props> = ({ history, onSelectionFromHistory, onClearHistory }) => {
+export const HistoryList: React.FC<Props> = ({ history, onSelectionFromHistory }) => {
   const onHistoryLinkClick = useCallback(
     (index: number) => {
       onSelectionFromHistory(index);
@@ -28,14 +25,6 @@ export const HistoryList: React.FC<Props> = ({ history, onSelectionFromHistory, 
 
   return history?.length ? (
     <>
-      <div className="api-client-sidebar-header">
-        {history?.length ? (
-          <RQButton type="transparent" size="small" onClick={onClearHistory} icon={<ClearOutlined />}>
-            Clear history
-          </RQButton>
-        ) : null}
-      </div>
-
       <Timeline reverse className="api-history-list" mode="left">
         <Timeline.Item key="end" color="gray">
           <div className="api-history-row">
